requirements.txt
streamlit==1.25.0
sentence_transformers==2.2.2
torch==2.0.1
Pillow==9.4.0